backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
autopep8==2.0.2
    # via -r requirements.in
boto3==1.26.117
    # via -r requirements.in
botocore==1.29.117
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
colorama==0.4.6
    # via pytest
coverage==7.2.3
    # via -r requirements.in
cryptography==40.0.2
    # via python-jose
django==4.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.14.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
exceptiongroup==1.1.1
    # via pytest
filelock==3.12.0
    # via tldextract
idna==3.4
    # via
    #   requests
    #   tldextract
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==23.1
    # via pytest
pillow==9.5.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycodestyle==2.10.0
    # via autopep8
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytest==7.3.1
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
python-jose[cryptography]==3.3.0
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
requests==2.28.2
    # via
    #   -r requirements.in
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
rsa==4.9
    # via python-jose
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
    #   requests-file
sqlparse==0.4.4
    # via django
tldextract==3.4.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   autopep8
    #   pytest
tzdata==2023.3
    # via django
urllib3==1.26.15
    # via
    #   botocore
    #   requests
